mudei muita coisa
