--- back/.env.example
@@ -1,1 +1,0 @@
-DATABASE_URL=""
