+++ back/README.md
@@ -1,1 +1,17 @@
-Aqui sera a pasta do back-end do projeto.
+# Back-End
+## _Estruta da Api_
+
+## Tech
+
+Este projeto utiliza as seguintes tecnologias:
+
+- Python - Uma linguagem de programação de alto nível, versátil e de fácil leitura
+- SQLAlchemy -  Uma biblioteca Python para interagir com bancos de dados relacionais usando mapeamento objeto-relacional (ORM).
+- FastApi -  Um framework rápido e moderno para criação de APIs em Python, com suporte a tipagem estática e documentação automática.
+
+
+### ORM
+
+ORM (Object-Relational Mapping) é uma biblioteca ou framework que mapeia objetos de uma aplicação para as tabelas de um banco de dados relacional, permitindo a persistência e recuperação dos dados de forma transparente, sem a necessidade de escrever consultas SQL diretamente. Ele automatiza tarefas como criação de tabelas, mapeamento de tipos de dados e execução de consultas, facilitando a interação entre a aplicação e o banco de dados.
+
+Quando um objeto é persistido em um banco de dados por meio de um ORM, o ORM analisa a estrutura do objeto e suas propriedades para determinar como mapeá-lo em uma tabela de banco de dados. Ele cria consultas SQL automaticamente para executar operações como inserção, atualização ou exclusão de dados. O ORM também gerencia o carregamento e a atualização dos objetos, mantendo-os sincronizados com o estado do banco de dados. Isso é feito por meio de uma camada de abstração que traduz as operações de banco de dados em chamadas de métodos de objeto, permitindo que os desenvolvedores interajam com o banco de dados usando linguagens de programação orientadas a objetos, em vez de escrever consultas SQL diretamente.
